perf(decode): route NVFP4 lm_head through batched M<=16 GEMV
The batched-decode lm_head ran the tiled w4a16_gemm over eff_n rows; the C=8
component profile put the vocab-projection GEMM at ~14% of the decode step,
reading the ~254 MB vocab weight in an uncoalesced pattern (~13-25% of peak BW).
Route it through the existing w4a16_gemv_batch16 kernel (M<=16, reads the weight
once for all rows) behind ATLAS_LMHEAD_BATCH_GEMV.

Measured C=8 (Qwen3.6-35B-A3B-NVFP4, GB10, Marlin MoE, graphs-on, clocks locked):
185.0 -> 206.0 tok/s (+11.4%), greedy byte-identical across a 5-prompt set
(the batch-GEMV reduction order differs from the tiled MMA, but both round to
the same bf16 logits, so greedy argmax is unchanged here).

Default OFF (opt-in) pending a longer soak; try_kernel falls back to the tiled
path on images without the kernel; eff_n>16 keeps the tiled path.

Also hoist the final-norm + lm-head + MS_PROFILE block out of decode_a2.rs into
decode_a2_head.rs (single call site, launch order unchanged) to keep decode_a2.rs
under the 500 LoC cap.
