+++ crates/spark-model/src/model/trait_impl/decode_a2.rs
@@ -389,82 +389,9 @@ impl TransformerModel {
                 None
             };
 
-            // Final norm [padded_n, H]
-            let normed = self.buffers.norm_output();
-            ops::rms_norm(
-                self.gpu.as_ref(),
-                self.rms_norm_kernel,
-                hidden,
-                &self.final_norm,
-                normed,
-                padded_n as u32,
-                h as u32,
-                self.config.rms_norm_eps as f32,
-                stream,
+            self.decode_final_norm_and_head(
+                hidden, n, padded_n, h, bf16, ssm_us, attn_us, lmhead_t0, stream,
             )?;
-
-            // LM head: ONE batched [padded_n, vocab] GEMM so the ~254 MB
-            // vocab weight is read ONCE per step instead of once per sequence
-            // (the per-row GEMV loop re-read it N times — a major C>=2 cost:
-            // ~N×254 MB/step). nvfp4/dense are batched here; FP8 single-scale
-            // keeps the per-row path (no batched single-scale FP8 GEMM handle
-            // on the model, and Holo's lm_head is NVFP4 anyway).
-            let logits = self.buffers.logits();
-            let v = self.config.vocab_size;
-            if let Some(ref fp8) = self.lm_head_fp8 {
-                for i in 0..padded_n {
-                    ops::dense_gemv_fp8w(
-                        self.gpu.as_ref(),
-                        self.dense_gemv_fp8w_kernel,
-                        normed.offset(i * h * bf16),
-                        fp8,
-                        logits.offset(i * v * bf16),
-                        v as u32,
-                        h as u32,
-                        stream,
-                    )?;
-                }
-            } else if let Some(ref nvfp4) = self.lm_head_nvfp4 {
-                ops::w4a16_gemm(
-                    self.gpu.as_ref(),
-                    self.w4a16_gemm_kernel,
-                    normed,
-                    nvfp4,
-                    logits,
-                    padded_n as u32,
-                    v as u32,
-                    h as u32,
-                    stream,
-                )?;
-            } else {
-                ops::dense_gemm(
-                    self.gpu.as_ref(),
-                    self.dense_gemm_kernel,
-                    normed,
-                    &self.lm_head_weight,
-                    logits,
-                    padded_n as u32,
-                    v as u32,
-                    h as u32,
-                    stream,
-                )?;
-            }
-            if let Some(t0) = lmhead_t0 {
-                self.gpu.synchronize(stream).ok();
-                let head_us = t0.elapsed().as_micros();
-                let total = ssm_us + attn_us + head_us;
-                tracing::info!(
-                    "ATLAS_MS_PROFILE n={n} padded_n={padded_n}: total={}us  ssm={}us({}L)  attn={}us({}L)  head={}us  [per-tok {:.2}ms]",
-                    total,
-                    ssm_us,
-                    self.config.num_ssm_layers(),
-                    attn_us,
-                    self.layers.len() - self.config.num_ssm_layers(),
-                    head_us,
-                    total as f64 / 1000.0 / padded_n as f64,
-                );
-            }
-
             if use_graphs {
                 let graph = self.gpu.end_capture(stream)?;
                 if graph.0 != 0 {
